chore: VendingMachineController -> ...Service, VendingMachine -> VendingMachineController로 이름 변경
