+++ src/main/java/vendingmachine/controller/VendingMachineController.java
@@ -1,78 +1,58 @@
 package vendingmachine.controller;
 
-import static vendingmachine.Constant.*;
-
 import java.util.ArrayList;
 
-import vendingmachine.domain.Price;
-import vendingmachine.domain.Product;
-import vendingmachine.repository.CoinRepository;
-import vendingmachine.repository.ProductRepository;
-import vendingmachine.utils.CoinGenerator;
-import vendingmachine.utils.ProductValidator;
+import vendingmachine.service.VendingMachineService;
+import vendingmachine.utils.StringUtil;
 import vendingmachine.view.InputView;
 import vendingmachine.view.OutputView;
 
 public class VendingMachineController {
-    private CoinRepository coinRepository;
-    private ProductRepository productRepository = new ProductRepository();
-    private Price userMoney;
+    private boolean isOperate = true;
+    private VendingMachineService vendingMachineService;
 
-    private VendingMachineController(int initializeMoney) {
-        putInitialAmount(initializeMoney);
+    public void start() {
+        initializeByAdmin();
+        useClient();
+        vendingMachineService.giveChange();
     }
 
-    public static VendingMachineController makeVendingMachineHasMoney(int initializeMoney) {
-        return new VendingMachineController(initializeMoney);
+    private void useClient() {
+        try {
+            String userMoneyInput = InputView.inputUserMoney();
+            vendingMachineService.putUserMoney(userMoneyInput);
+            while (isOperate) {
+                String productName = InputView.inputBuyingProduct();
+                isOperate = vendingMachineService.sellProduct(productName);
+            }
+        } catch (IllegalArgumentException e) {
+            OutputView.showErrorMessage(e);
+            useClient();
+        }
     }
 
-    public void putInitialAmount(int inputMoney) { //금액 검증 로직 만들었으니 그걸 사용하기.
-        if (inputMoney < ZERO) {
-            throw new IllegalArgumentException("0 이상의 금액을 입력해주세요.");
-        }
-        if (inputMoney % MINIMUM_COIN_VALUE != 0) {
-            throw new IllegalArgumentException("해당 금액은 동전으로 만들 수 없는 단위의 숫자입니다.");
-        }
-        coinRepository = new CoinRepository(CoinGenerator.makeCoins(inputMoney));
-        // System.out.println(coinRepository);
-        OutputView.showAllCoinsMachineHave(coinRepository);
-
+    private void initializeByAdmin() {
+        makeCoinsUsingEnteredAmount();
+        putProductInVendingMachine();
     }
 
-    public void putProducts(ArrayList<String> productsInfo) {
-        productRepository.clear();
-        for (String productInfoInput : productsInfo) {
-            Product product = ProductValidator.validateForm(productInfoInput);
-            productRepository.add(product);
+    private void putProductInVendingMachine() {
+        try {
+            ArrayList<String> productsInfo = StringUtil.splitUsingSemiColon(InputView.inputProductsInfo());
+            vendingMachineService.putProducts(productsInfo);
+        } catch (IllegalArgumentException e) {
+            OutputView.showErrorMessage(e);
+            putProductInVendingMachine();
         }
-        if (productRepository.hasNoQuantity()) {
-            throw new IllegalArgumentException("자판기에는 최소 한 개의 물건은 들어가야 합니다.");
-        }
-        // productRepository.showProductRepository();
     }
 
-    public void putUserMoney(String userMoneyInput) {
-        userMoney = new Price(userMoneyInput);
-        if (productRepository.cantBuyBecauseOfNoMoney(userMoney)) {
-            throw new IllegalArgumentException("투입한 금액으로 살 수 있는 물건은 존재하지 않습니다.");
+    private void makeCoinsUsingEnteredAmount() {
+        try {
+            int initializeMoney = StringUtil.parseStringToInt(InputView.inputInitialAmount());
+            vendingMachineService = VendingMachineService.makeVendingMachineHasMoney(initializeMoney);
+        } catch (IllegalArgumentException e) {
+            OutputView.showErrorMessage(e);
+            makeCoinsUsingEnteredAmount();
         }
-        OutputView.showUserMoney(userMoney);
-    }
-
-    public boolean sellProduct(String productName) {
-        if (!productRepository.has(productName)) {
-            throw new IllegalArgumentException("해당 상품은 존재하지 않습니다.");
-        }
-        int productPrice = productRepository.takeout(productName);
-        userMoney.use(productPrice);
-        OutputView.showUserMoney(userMoney);
-        if (productRepository.cantBuyBecauseOfNoMoney(userMoney) || productRepository.hasNoQuantity()) {
-            return false;
-        }
-        return true;
-    }
-
-    public void giveChange() {
-        OutputView.showChange(coinRepository.giveChange(userMoney));
     }
 }
